Implemented malloc and free

+++ src/kstd.hh
@@ -1,10 +1,12 @@
 #pragma once
 
+#include "common.hh"
+
 namespace kstd {
     struct ___initialization_params {
         // Heap related information
         void* heap_start;
-        uint64 heap_size;
+        uint32 heap_size;
     };
 
     void ___initialize(const ___initialization_params& params);
